requirements.txt
Flask==3.1.0
boto3==1.35.0
python-dotenv==1.0.0 